fix(frontend): address #9972 review — drop dead nav flag entry, extract vision types, UX nits (#9890)
